+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Head
 
+- Removed: `no-confusing-arrow` because of the deprecation of formatting rules.
 - Changed: replace deprecated `padding-line-between-statements` with `@stylistic/js/padding-line-between-statements`.
 - Changed: bump `eslint-plugin-jest` from v27 to v28.
 - Changed: bump `eslint-plugin-n` from v16 to v17.